requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
google-auth
google-cloud-storage
jsonschema
pytest
PyYAML>=6.0
google-generativeai>=0.3.0